requirements.txt
ultralytics
opencv-python
torch
scikit-learn
rasterio
piexif
tqdm
pillow
